refactor: ui pass event for reply
This one is interesting... replying involves the view and the compose pane.

Here's how we handle it"
* reply button in view pane dispatches an event.
* we listen to it in the ui setup
* from there we trigger the reply event of the compose pane.
* we hand the message we want to reply to along
* compose pane knows how to reply to a given message.

In theory one pane could listen to the other but that would require
that it knows about the other. I currently prefer turning the ui.js into
some kind of controller and keeping the views for the different panes
ignorant of each other.